refactor PreJSRewriter to PreJSOptimize
